Implemented SMS ACtions lock & childmode

+++ _parentControl/app/src/main/java/com/praskum/parentcontrol/AlarmReceiver.java
@@ -4,7 +4,6 @@ import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
 import android.database.Cursor;
-import android.icu.util.UniversalTimeScale;
 import android.media.AudioManager;
 import android.net.wifi.WifiManager;
 import android.provider.Settings;
@@ -16,60 +15,26 @@ public class AlarmReceiver extends BroadcastReceiver {
     public void onReceive(Context context, Intent intent) {
         Log.i("AlarmReceiver", "Alarm Received");
 
-        int id = intent.getIntExtra("id", -1);
-        DatabaseHelper dbHelper = new DatabaseHelper(context);
-        Cursor c = dbHelper.ReadData(id);
+        if (intent.getAction().equalsIgnoreCase(Constants.TIMERINTENTACTION)) {
+            int id = intent.getIntExtra("id", -1);
 
-        AudioManager am;
-        am = (AudioManager) context.getSystemService(Context.AUDIO_SERVICE);
-
-        if (c.getCount() > 0) {
-            c.moveToFirst();
-            if (c.getInt(c.getColumnIndex("Wifi")) == 1) {
-                final WifiManager wifi = (WifiManager) context.getSystemService(Context.WIFI_SERVICE);
-                Log.i("AlarmReceiver", "Switching off wifi");
-                wifi.setWifiEnabled(false);
+            if (id == Constants.RESET_SCREEN_OFF_TIMEOUT_ALARMID) {
+                int screenOffTimeout = intent.getIntExtra("value", -1);
+                Settings.System.putInt(context.getContentResolver(), Settings.System.SCREEN_OFF_TIMEOUT, screenOffTimeout);
             }
+            else {
+                DatabaseActionHelper dbHelper = new DatabaseActionHelper(context);
+                Cursor c = dbHelper.ReadData(id);
+                if (c != null && c.getCount() > 0) {
+                    ActionRunner actionRunner = new ActionRunner();
+                    actionRunner.ApplyActions(context, c);
+                }
 
-            if (c.getInt(c.getColumnIndex("Silent")) == 1) {
-                Log.i("AlarmReceiver", "Silent mode turned on");
-                am.setRingerMode(AudioManager.RINGER_MODE_SILENT);
+                boolean isDeleted = dbHelper.DeleteData(id);
+                if (!isDeleted) {
+                    Toast.makeText(context, "Failed to delete the record", Toast.LENGTH_LONG).show();
+                }
             }
-
-            if (c.getInt(c.getColumnIndex("SwitchToHomeScreen")) == 1) {
-                Intent startMain = new Intent(Intent.ACTION_MAIN);
-                startMain.addCategory(Intent.CATEGORY_HOME);
-                startMain.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-                Log.i("AlarmReceiver", "Switching to home screen");
-                context.startActivity(startMain);
-            }
-
-            int currentVolume = am.getStreamVolume(AudioManager.STREAM_MUSIC);
-            int newVolume = c.getInt(c.getColumnIndex("MediaVolume"));
-
-            int delta = Math.abs(newVolume - currentVolume);
-            int v = currentVolume > newVolume ? AudioManager.ADJUST_LOWER : AudioManager.ADJUST_RAISE;
-
-            while (delta > 0) {
-                delta--;
-                am.adjustStreamVolume(AudioManager.STREAM_MUSIC, v, AudioManager.FLAG_PLAY_SOUND);
-            }
-
-            int newBrightness = c.getInt(c.getColumnIndex("Brightness"));
-            if (newBrightness >= 0) {
-                Settings.System.putInt(context.getContentResolver(), Settings.System.SCREEN_BRIGHTNESS, newBrightness);
-            }
-
-            if (c.getInt(c.getColumnIndex("ScreenOff")) == 1) {
-                int defaultTurnOffTime =  Settings.System.getInt(context.getContentResolver(),Settings.System.SCREEN_OFF_TIMEOUT, 60000);
-                Values.ScreenOffTimeout = defaultTurnOffTime;
-                Settings.System.putInt(context.getContentResolver(), Settings.System.SCREEN_OFF_TIMEOUT, 10);
-            }
-        }
-
-        boolean isDeleted = dbHelper.DeleteData(id);
-        if (!isDeleted) {
-            Toast.makeText(context, "Failed to delete the record", Toast.LENGTH_LONG).show();
         }
     }
 }
